+++ app/src/main/java/com/rendox/grocerygenius/feature/add_grocery/AddGroceryBottomSheetContent.kt
@@ -291,7 +291,7 @@ private fun SearchResultsPreview() {
                 category = Category(
                     id = index.toString(),
                     name = "Category$index",
-                    sortingPriority = index,
+                    sortingPriority = index.toLong(),
                 ),
             )
         }
